+++ htmf/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Output unformatted HTML (without newlines or indentation) by default. Add the "pretty-print" feature along with a new method, `to_html_pretty`, for opt-in formatted HTML output.
 - Don't render a closing tag for [void elements](https://developer.mozilla.org/en-US/docs/Glossary/Void_element).
+- Replace implementation of `IntoElements for Vec<Into<Element>>` with implementation of `From<Vec<Into<Element>>> for Element` to allow easily creating fragments in a list of children
 
 # Changelog